fix: move AICategoryService to HistoryService
